Split ScriptObject.h into subheaders
