+++ core/components/scriptable/TransformComponent.h
@@ -4,9 +4,9 @@
 #pragma once
 
 #include "core/components/ScriptableComponent.h"
-#include "core/scripting/object/ScriptObject.h"
 #include "core/world/Entity.h"
 #include "lib/include/glm_headers.h"
+#include "lib/include/wasm_headers.h"
 #include "types/assets/PrefabAsset_generated.h"
 #include "types/protocol/ProtocolTypes.h"
 #include "types/protocol/TransformComponent_generated.h"
